requirements.txt
annotated-types==0.7.0
anyio==4.4.0
argcomplete==3.4.0
attrs==24.2.0
certifi==2024.7.4
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
graphql-core==3.2.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.8
mypy==1.11.1
mypy-extensions==1.0.0
neovim==0.3.1
packaging==24.1
pipx==1.6.0
platformdirs==4.2.2
pluggy==1.5.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pynvim==0.5.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
strawberry-graphql==0.237.2
typer==0.12.3
typing_extensions==4.12.2
userpath==1.9.2
uvicorn==0.30.4
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
